First upload of currently used trackers
eXtoplasmVR was created because of my love for the SlimeX cases but I wanted to upgrade my trackers to BNO085 IMUs and wanted to include a PCB to reduce issues I was having with reliability of soldered wires. I modelled the tracker cases by eyeballing the original SlimeX cases after designing the PCB to make it fit and kept as much of the design language of the original cases as I could as I loved the compactness and comfort that these cases gave me whilst I used the originals with BMI160 IMU's.

To the SlimeX team - thank you for inspiring me to brush off my CAD skills - I know there is a lot of stuff wrong with these trackers and a lot of stuff I would like to change and refine and improve but for now, I think I am happy enough with them that I would like to release these in to the wild in the hopes that they help someone or that someone else can find them useful.

Also after printing probably about 100 revisions over the months and refining them, my printer needs a rest. She has done well for being 10 years old nearly lol.

Still to do:

- Make a mattery protector (I currently just cover the back of the PCB facing the battery with a few layers of blue painters tape I use on my buildplate for my printer.